misc/HtmlTestSuite/results/test-0741.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mailto:mstensho@chromium.org" rel="author" title="Morten Stenshorne"/>
<link href="https://www.w3.org/TR/CSS22/visuren.html#propdef-left" rel="help"/>
<link href="https://www.w3.org/TR/CSS22/visuren.html#relative-positioning" rel="help"/>
<link href="https://www.w3.org/TR/CSS22/visuren.html#anonymous-block-level" rel="help"/>
<link href="https://bugs.chromium.org/p/chromium/issues/detail?id=1002485" rel="help"/>
</head><body><p>There should be no red, and no scrollbar.</p>
<div id="container" style="overflow:auto; width:500px; background:red;">
  <div style="padding-right:90%; background:yellow;">
    <div style="position:relative; left:900%; background:cyan;">
      <div/>
       
    </div>
  </div>
</div>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
  container.scrollLeft = 123456;
  test(()=&gt; {
      assert_equals(container.scrollLeft, 0);
  }, "Left percentage resolved correctly for overflow contribution");
</script>
</body></html>
